feat: adicionar poster, nome do genero e refatorar o serviço de recomendação de filmes

+++ src/main/java/com/ezequiel/moviematch/api/record/genre/GenreRecord.java
@@ -1,4 +1,4 @@
 package com.ezequiel.moviematch.api.record.genre;
 
-public record GenreRecord(Long id, String name) {
+public record GenreRecord(Integer id, String name) {
 }
